Add retries to the `land` command
This is an attempt to fix https://github.com/getcord/spr/issues/108

It seems that by changing the base branch to master unconditionally in
the UI or using `gh` the land succeeds in a large repo that has branch
protection on. This is an attempt at doing that directly in the land
command. We also add retries in the case that there is a propagation
delay on GitHub's side.